\name{connection-class}
\docType{class}
\alias{connection-class}
\alias{file-class}

\title{Virtual S4 classes for method dispatching}

\description{
  Virtual S4 classes for method dispatching}

\details{
These two classes are used to simulate the S3 class 'file' extending
class 'connection' for method dispatching. They can't be instantiated!  
}
			
\author{Florian Hahne}

\keyword{classes}

