\name{eset133a}
\alias{eset133a}
\alias{expressionSet133a}
\docType{data}
\title{A small dataset for testing}
\description{
	An artificial Affymetrix hgu133a dataset, with one 
covariate 'cov1'.
}
\usage{data(eset133a)
       data(expressionSet133a)
}
\format{
 The data are artifical. There are 6 cases labeled 1 to
 6 and and 22283 genes as in an Affymetrix U133a chips. 
 There is one covariate (factor) whose values are "type 1"
 for the first 3 samples and "type 2" for the last 3 samples.
}
\examples{
data(eset133a)
data(expressionSet133a)
}
\keyword{datasets}
